Refactor to decouple services
